+++ src/frontend/Player/createHlsPlayer.spec.ts
@@ -1,6 +1,6 @@
 import { createHlsPlayer } from './createHlsPlayer';
 
-import { Video } from '../types/tracks';
+import { VideoUrls } from '../types/tracks';
 
 jest.mock('hls.js', () => {
   return jest.fn(() => ({
@@ -16,6 +16,8 @@ describe('createHlsPlayer', () => {
       manifests: {
         hls: 'https://example.com/hls.m3u8',
       },
+      mp4: {},
+      thumbnails: {},
     },
   };
 
@@ -24,7 +26,7 @@ describe('createHlsPlayer', () => {
   });
 
   it('creates and configure hls.js player', () => {
-    const player = createHlsPlayer(video as Video, 'ref' as any);
+    const player = createHlsPlayer(video.urls as VideoUrls, 'ref' as any);
     expect(player.attachMedia).toHaveBeenCalledWith('ref');
     expect(player.loadSource).toHaveBeenCalledWith(
       'https://example.com/hls.m3u8',