+++ src/frontend/Player/createHlsPlayer.ts
@@ -1,16 +1,15 @@
 import Hls from 'hls.js';
-
-import { Video } from '../types/tracks';
+import { VideoUrls } from '../types/tracks';
 
 export const createHlsPlayer = (
-  video: Video,
+  urls: VideoUrls,
   videoNode: HTMLVideoElement,
 ): Hls => {
   const hls = new Hls({
     liveDurationInfinity: true,
     liveSyncDuration: 1,
   });
-  hls.loadSource(video.urls.manifests.hls);
+  hls.loadSource(urls.manifests.hls);
   hls.attachMedia(videoNode);
 
   return hls;